[CHANGE] move data checker to init_pipeline
