+++ application/gpytorch_pipeline.py
@@ -2,7 +2,7 @@ import gpytorch
 import torch
 from adapters.gpytorch.gp_model import MyExactGP, MyApproximateGP
 from gpytorch.mlls import ExactMarginalLogLikelihood, InducingPointKernelAddedLossTerm, VariationalELBO, GammaRobustVariationalELBO
-from application.init_pipeline import init_pipeline, get_numpy_features
+from application.init_pipeline import init_pipeline, locate_invalid_data, validate_data
 from domain.env import USE_DUMMY_DATA, MODELDIR, MEAN_FUNC, KERNEL_TYPE, KERNEL_STRUCTURE, DATA_SLICE_AMOUNT, SWS, Y, POLY_DEGREE
 from domain.metrics import get_metrics, gaussian_log_likelihood
 import numpy as np
@@ -27,24 +27,6 @@ def get_data(get_ds=False):
         return (ds, X_train, X_test, y_train, y_test, feature_names)
     else:
         return (X_train, X_test, y_train, y_test, feature_names)
-
-def locate_invalid_data(data):
-    if isinstance(data, torch.Tensor):
-        isnan = torch.isnan(data)
-        isinf = torch.isinf(data)
-    elif isinstance(data, np.ndarray):
-        isnan = np.isnan(data)
-        isinf = np.isinf(data)
-    
-    invalid_data_locs = isnan | isinf
-    return invalid_data_locs
-
-def validate_data(*args):
-    for arg in args:
-        if not torch.isfinite(arg).all():
-            print(locate_invalid_data(arg))
-            raise ValueError("Data contains NaN or inf values.")
-    print(f"data is fine.")
 
 def choose_model(model="exact", data=None):
     if not data:
